L4: move part of the code into alphabet.c/h, add to build.sh
